run as tank
